ml_service/requirements.txt
grpcio==1.71.2
grpcio-tools==1.71.2
grpcio-reflection==1.71.2
numpy==2.2.0
pandas==2.2.3
torch==2.9.1
xgboost==2.1.3
scikit-learn==1.6.0
transformers==4.47.1
tensorflow==2.20.0
ta==0.11.0
python-dotenv==1.0.1
loguru==0.7.3